project/backend/requirements.txt
# Requirements for NeuroCore RAG Backend

# Core LLM & NLP
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
langchain>=0.1.0

# Vector Database
faiss-cpu>=1.7.4

# Document Processing
PyPDF2>=3.0.0

# API (Optional)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Utilities
numpy>=1.24.0
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
